+++ pysat/instruments/timed_see.py
@@ -38,9 +38,9 @@ Warnings
 
 from __future__ import print_function
 from __future__ import absolute_import
+import datetime as dt
 import functools
 
-import pysat
 from pysat.instruments.methods import nasa_cdaweb as cdw
 
 # include basic instrument info
@@ -48,7 +48,7 @@ platform = 'timed'
 name = 'see'
 tags = {'': ''}
 sat_ids = {'': ['']}
-_test_dates = {'': {'': pysat.datetime(2009, 1, 1)}}
+_test_dates = {'': {'': dt.datetime(2009, 1, 1)}}
 
 
 # support list files routine