+++ pysat/_instrument.py
@@ -349,6 +349,7 @@ class Instrument(object):
                 # default provided by instrument module
                 orbit_info = self.orbit_info
         self.orbits = _orbits.Orbits(self, **orbit_info)
+        self.orbit_info = orbit_info
 
         # Create empty placeholder for meta translation table
         # gives information about how to label metadata for netcdf export
@@ -909,7 +910,7 @@ class Instrument(object):
 
         output_str += '\nOrbit Settings' + '\n'
         output_str += '--------------' + '\n'
-        if self.orbit_info is None:
+        if self.orbits.orbit_index is None:
             output_str += 'Orbit properties not set.\n'
         else:
             output_str += 'Orbit Kind: ' + self.orbit_info['kind'] + '\n'