+++ CHANGELOG.md
@@ -26,7 +26,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Bug Fix
   - Fixed custom instrument attribute persistence upon load
 - Maintenance
-  - Specify dtype for pandas.Series(None) for forward compatibility
+  - Specify dtype for empty pandas.Series for forward compatibility
 
 ## [2.2.0] - 2020-2-29
 - New Features