+++ CHANGELOG.md
@@ -57,6 +57,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - Added small time offsets (< 1s) to ensure COSMIC files and data have unique times
   - Updates to Travis CI environment
   - Removed `inplace` use in xarray `assign` function, which is no longer allowed
+  - Fixed output of orbit_info during print(inst)
 
 
 ## [2.1.0] - 2019-11-18