+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [2.2.0] - 2020-2-29
+## [2.2.0] - 2020-4-27
 - New Features
    - Decreased time to load COSMIC GPS data by about 50%
    - Added DE2 Langmuir Probe, NACS, RPA, and WATS instruments
@@ -36,6 +36,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - Removed `inplace` use in xarray `assign` function, which is no longer allowed
   - Removed old code and incorrect comments from F10.7 support
   - Updated use of numpy.linspace to be compatible with numpy 1.18.
+  - Fixed output of orbit_info during print(inst)
 
 
 ## [2.1.0] - 2019-11-18